Add parallelisation to kaem
Enabled by the _begin_parallel and _end_parallel commands. For example,

```
_begin_parallel 4
cc a.c
cc b.c
[...]
_end_parallel
```

will execute up to 4 of the cc commands in parallel at a time.

Note that this is lazy parallelisation; all commands are run at the
_end_parallel command, in a non-deterministic order. As such, builtins
are not permitted to be run in a _begin_parallel/_end_parallel block.